+++ changelog.md
@@ -1,5 +1,8 @@
 ## Change logs
 
+#v1.3.7
+* Added "compatible Magisk-mirroring" message for incompatible Magisk variants
+
 #v1.3.6
 * Tuned the USB period size for SDM845 devices (2500 usec to 2250 usec)
 * Tuned the USB period size for other devices (to 2250 usec)